requirements.txt
requests>=2.31.0 
google-adk>=1.0.0
flask>=2.3.0
aiohttp>=3.8.0 